requirements.txt
pyyaml~=6.0
tensorflow~=2.12.0
paramiko~=2.8.1
pandas~=1.5.3
mlflow~=2.3.2
pyinstrument~=4.4.0